fixed testcases for basic IO
