+++ XiansAi.Lib.csproj
@@ -22,6 +22,7 @@
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.3" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
     <PackageReference Include="OpenAI" Version="2.1.0" />
+    <PackageReference Include="Moq" Version="4.20.70" />
     <PackageReference Include="Temporalio" Version="1.5.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.3" />
     <PackageReference Include="Microsoft.Extensions.Options" Version="9.0.3" />
